+++ tests/output_test.py
@@ -32,11 +32,12 @@ sys.path.insert(1, modules_path)
 from f4enix.input.libmanager import LibManager
 from jade.configuration import Configuration
 import jade.output as output
+import pandas as pd
 from jade.expoutput import SpectrumOutput
-from jade.sphereoutput import SphereOutput
+import jade.sphereoutput as sout
 from jade.configuration import Configuration
 from jade.__version__ import __version__
-from f4enix.output.MCNPoutput import MCNPoutput
+from jade.output import MCNPoutput
 
 # Files
 OUTP_SDDR = os.path.join(
@@ -45,6 +46,29 @@ OUTP_SDDR = os.path.join(
 OUTM_SDDR = os.path.join(
     cp, "TestFiles", "sphereoutput", "SphereSDDR_11023_Na-23_102_m"
 )
+
+
+class TestSphereSDDRMCNPoutput:
+
+    out = sout.SphereSDDRMCNPoutput(OUTM_SDDR, OUTP_SDDR)
+
+    def test_get_single_excel_data(self):
+        vals, errors = self.out.get_single_excel_data()
+        assert isinstance(vals, pd.Series)
+        assert isinstance(errors, pd.Series)
+        assert len(vals) == 23
+        assert len(errors) == 23
+
+
+class TestMCNPoutput:
+    def test_mcnpoutput(self):
+        out = MCNPoutput(OUTM_SDDR, OUTP_SDDR)
+        t4 = out.tallydata[4]
+        t2 = out.tallydata[2]
+        assert list(t4.columns) == ["Cells", "Segments", "Value", "Error"]
+        assert len(t4) == 1
+        assert len(t2) == 176
+        assert list(t2.columns) == ["Energy", "Value", "Error"]
 
 
 class MockSession:
