+++ tests/sphereoutput_test.py
@@ -41,7 +41,7 @@ from jade.configuration import Configuration
 from f4enix.input.libmanager import LibManager
 from jade.status import Status
 from jade.__version__ import __version__
-
+from jade.output import MCNPoutput
 import jade.sphereoutput as sout
 
 
@@ -205,32 +205,3 @@ class TestSphereSDDRoutput:
         # print(errors)
         # print(stat_checks)
         # print(results.columns)
-
-
-cp = os.path.dirname(os.path.abspath(__file__))
-# Files
-OUTP_SDDR = os.path.join(cp, "resources", "output", "SphereSDDR_11023_Na-23_102_o")
-OUTM_SDDR = os.path.join(cp, "resources", "output", "SphereSDDR_11023_Na-23_102_m")
-
-
-class TestSphereSDDRMCNPoutput:
-
-    out = SphereSDDRMCNPoutput(OUTM_SDDR, OUTP_SDDR)
-
-    def test_get_single_excel_data(self):
-        vals, errors = self.out.get_single_excel_data()
-        assert isinstance(vals, pd.Series)
-        assert isinstance(errors, pd.Series)
-        assert len(vals) == 23
-        assert len(errors) == 23
-
-
-class TestMCNPoutput:
-    def test_mcnpoutput(self):
-        out = MCNPoutput(OUTM_SDDR, OUTP_SDDR)
-        t4 = out.tallydata[4]
-        t2 = out.tallydata[2]
-        assert list(t4.columns) == ["Cells", "Segments", "Value", "Error"]
-        assert len(t4) == 1
-        assert len(t2) == 176
-        assert list(t2.columns) == ["Energy", "Value", "Error"]
